requirements.txt
python-telegram-bot==20.7
gspread==5.11.3
oauth2client==4.1.3
pillow==10.3.0
whisper-openai==0.2.2
python-dotenv==1.0.1
pytesseract==0.3.10
